audiobook_project/requirements.txt
django
djangorestframework
django-cors-headers
google-generativeai
edge-tts
pydub
python-dotenv
tqdm